Segnos, Voltas, TempoVariances and Repeats with JUnit tests
New functionality added to handles Segnos (all covered by JUnit tests).
All produce MusicXML readable by MuseScore, and all but two readable by
Sibelius - which doesn't support DalSegno and DaCapo.

Voltas implemented and tested on single stave scores - although the code
is written to handle voltas on all staves - further tests need to be
written for multiple staves.

Repeats - NWC has a repeat capability on light-light bar lines (coded in
NWC as two light-light bar lines with double dots with a number in
brackets above the second bar line indicating the sequence of bars is to
be repeated the number of times in the brackets.  MusicXML supports bar
repeat sections - but with a pair of heavy-light light-heavy bar lines
with double dots.  So the code translates NWC light-light into
heavy-light/light-heavy when it's a repeat bar - and the resulting
MusicXML works in MuseScore, but not in Sibelius - which has a bug - the
repeat can be coded in the Sibelius GUI but when Sibelius exports to
MusicXML even though the XMLis right, when imported back into Sibelius
(or from the NWCTXT conversion from this programme) Sibelius displays
correctly (the (n) repeat text in NWC is represented by "Play n times"
in MusicXML - but the repeats do not work.  Presumably Sibelius will fix
this bug in time.

Also tempo variances (e.g. rall. rit. etc).  In NWC this set includes
both breath-mark and fermata.

All tests use a variation on one of the compare files methods in IOUtils
- which has been modified by ignore the encoding-date tag line - as this
has the date the XML is written - so needs to be excluded from the file
diff for the tests to be successful.