Fix pre-segnos voltas repeats regression test
In my previous checked in code I fixed a problem whereby with a grand
stave, the key/time signatures were added twice (as Nwc2MusicXML
processed each stave in the NWC file, it added the key/time signatures
to the output, meaning on a grand stave you end up with both twcie).
Sibelius loads the output file, but MuseScore, with its more faithful
interpreation of MusicXML objects.  Although I fixed the problem in the
code, I appear not to have fixed the test.  I've now updated the
reference file, as the output is correct.
I've also commented out a System.err.println statement in the
Nwc2MusicXML.java file, which was a debugging line that had been left
uncommented (it prints every Wedge to System.err).