Add support for 2 voices on stave (including tests)
NWC implements two voices on a stave inconsistently.  It uses stem
up/down on the first note of a chord, but the note that is given stem
up/down is the first note input into NWC irrespecitve of whether it is
voice 1 or voice 2.  So to correctly determine the voice of a note you
have to look at both stem direction and the Pos/Pos2 value.  Without
doing this (which nwc2musicxml did before this fix) the same voice flips
from up stem to down stem in NWCTXT files and so was being written into
XML inconsistently.
The fix is to read both stem direction and Pos/Pos2 and determine the
voice from that.
There is a test TwoVoiceChords for the two voices being handled
correctly; and an additional regression test of a long piece of 1 voice
music (a Beethoven fantasy) to confirm that the new code has no impact
on single voice music.
There's also a change to the reference file NWCReallyCompleteExample
(the test file that has one of every notation NWC can produce) as this
has two voices and was previously rendered incorrectly.